switched over to jest and new testing structure
