+++ .eslintignore
@@ -1,5 +1,5 @@
 **/*{.,-}min.js
 /public
-/config
+/config/**
 registerServiceWorker.js
 /src
