+++ package.json
@@ -41,12 +41,13 @@
     "autoprefixer": "7.1.6",
     "babel-core": "6.26.0",
     "babel-eslint": "7.2.3",
-    "babel-jest": "20.0.3",
+    "babel-jest": "23.4.2",
     "babel-loader": "7.1.2",
     "babel-preset-react-app": "^3.1.1",
     "babel-preset-stage-2": "6.24.1",
     "babel-runtime": "6.26.0",
     "case-sensitive-paths-webpack-plugin": "2.1.1",
+    "chai": "4.1.2",
     "chalk": "1.1.3",
     "css-loader": "0.28.7",
     "enzyme": "3.3.0",
@@ -66,7 +67,7 @@
     "flow-bin": "^0.78.0",
     "fs-extra": "3.0.1",
     "html-webpack-plugin": "2.29.0",
-    "jest": "20.0.4",
+    "jest": "23.4.2",
     "object-assign": "4.1.1",
     "postcss-flexbugs-fixes": "3.2.0",
     "postcss-loader": "2.0.8",
@@ -74,6 +75,8 @@
     "raf": "3.4.0",
     "react-dev-utils": "^5.0.1",
     "resolve": "1.6.0",
+    "sinon": "6.1.4",
+    "sinon-chai": "3.2.0",
     "style-loader": "0.19.0",
     "sw-precache-webpack-plugin": "0.11.4",
     "url-loader": "0.6.2",
@@ -83,11 +86,15 @@
     "whatwg-fetch": "2.0.3"
   },
   "jest": {
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx,mjs}"
-    ],
+    "setupTestFrameworkScriptFile": "./src/setupTest.js",
     "setupFiles": [
       "<rootDir>/config/polyfills.js"
+    ],
+    "globals": {
+      "ENV": "test"
+    },
+    "collectCoverageFrom": [
+      "src/**/*.{js,jsx,mjs}"
     ],
     "testMatch": [
       "<rootDir>/src/**/__tests__/**/*.{js,jsx,mjs}",
@@ -106,6 +113,12 @@
     "moduleNameMapper": {
       "^react-native$": "react-native-web"
     },
+    "moduleDirectories": [
+      "<rootDir>node_modules",
+      "<rootDir>src",
+      "<rootDir>__test__",
+      "<rootDir>common"
+    ],
     "moduleFileExtensions": [
       "web.js",
       "js",
