+++ src/layout/Auth/components/Auth.spec.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { expect } from 'chai';
+import { shallow } from 'enzyme';
+
+import Auth from './Auth.jsx';
+
+describe('Auth.jsx', () => {
+  it('Changes tabs', () => {
+      const authTest = shallow(<Auth authCancelled={() => {}} />);
+      authTest.instance().handleTabChange(1);
+      expect(authTest.state().index).to.equal(1);
+    });
+});
