+++ src/layout/Auth/components/Register.spec.jsx
