+++ src/pages/Settings/Settings.jsx
@@ -7,8 +7,8 @@ import Header from 'grommet/components/Header';
 import Button from 'grommet/components/Button';
 
 import RequireAuth from 'hoc/RequireAuth.jsx';
-import AccountContainer from '../routes/Account';
-import ProfileContainer from '../routes/Profile';
+import { AccountContainer } from './routes/Account';
+import { ProfileContainer } from './routes/Profile';
 
 class Settings extends React.Component {
   navigate = (url) => {
