+++ src/pages/Settings/index.js
@@ -1,1 +1,1 @@
-export { default as Settings } from './components/Settings.jsx';
+export { default as Settings } from './Settings.jsx';
