+++ src/pages/Settings/routes/Account/AccountContainer.jsx
@@ -3,8 +3,8 @@ import * as firebase from 'firebase';
 import { connect } from 'react-redux';
 import update from 'immutability-helper';
 
-import Account from './Account.jsx';
-import {updateUser} from 'redux/actions.js';
+import { Account } from './components';
+import { updateUser } from 'redux/actions.js';
 import logger from 'logger.js';
 
 class AccountContainer extends React.Component {
