+++ src/pages/Settings/routes/Account/components/index.js
@@ -1,0 +1,1 @@
+export { default as Account } from './Account.jsx';
