+++ src/pages/Settings/routes/Account/index.js
@@ -1,1 +1,1 @@
-export { default as AccountContainer } from './components/AccountContainer.jsx';
+export { default as AccountContainer } from './AccountContainer.jsx';
