+++ src/pages/Settings/routes/Profile/components/index.js
@@ -1,0 +1,1 @@
+export { default as Profile } from './Profile.jsx';
