+++ src/pages/Settings/routes/Profile/index.js
@@ -1,1 +1,1 @@
-export { default } from './components/ProfileContainer.jsx';
+export { default } from './ProfileContainer.jsx';
