--- test/components/Auth/index.js
@@ -1,3 +1,0 @@
-import testAuth from './test-auth.jsx';
-import testLogin from './test-login.jsx';
-import testRegister from './test-register.jsx';
