--- test/index.js
@@ -1,3 +1,0 @@
-import componentTests from './components';
-import routesTests from './routes';
-import serverTests from './server';
