--- test/routes/Settings/components/index.js
@@ -1,1 +1,0 @@
-import testSettings from './test-settings.jsx';
