--- test/routes/Settings/index.js
@@ -1,2 +1,0 @@
-import componentTests from './components';
-import routesTests from './routes';
