--- test/routes/Settings/routes/Account/components/index.js
@@ -1,1 +1,0 @@
-import testAccount from './test-account.jsx';
