--- test/routes/Settings/routes/Account/index.js
@@ -1,1 +1,0 @@
-import componentTests from './components';
