--- test/routes/Settings/routes/Profile/components/index.js
@@ -1,1 +1,0 @@
-import testProfile from './test-profile.jsx';
