--- test/routes/Settings/routes/Profile/index.js
@@ -1,1 +1,0 @@
-import componentTests from './components';
