--- test/routes/Settings/routes/index.js
@@ -1,2 +1,0 @@
-import accountTests from './Account';
-import profileTests from './Profile';
