--- test/server/index.js
@@ -1,4 +1,0 @@
-import testAuth from './test-auth.js';
-import testUser from './test-user.js';
-import testEvents from './test-events.js';
-import testReports from './test-reports.js';
